Move crl fixtures to conftest
